visualization: Exported constants and utility functions to separate files
Also reordered callbacks, fixed imports in the test files and added a mean value to run visualization
